+++ frontend/app/src/entities/nodes/object/ui/object-table/toolbar/actions/objects/delete-objects-modal.tsx
