+++ frontend/app/src/entities/nodes/object/ui/object-table/toolbar/actions/objects/toolbar-delete-action.tsx
@@ -2,7 +2,7 @@ import { ToolbarButton } from "@/entities/nodes/object/ui/object-table/toolbar/t
 import { NodeObject } from "@/entities/nodes/types";
 import { Icon } from "@iconify-icon/react";
 import React from "react";
-import { DeleteObjectsModal } from "./delete-object-modal";
+import { DeleteObjectsModal } from "./delete-objects-modal";
 
 export interface ToolbarDeleteObjectProps {
   selectedRows: Array<NodeObject>;
